requirements.txt
# Media Tools - Requirements
# ===========================
# Dependências Python para processamento de mídia
#
# Python 3.6+ é necessário
#
# Para instalar todas as dependências:
#   pip install -r requirements.txt
#
# Para instalar em ambiente virtual (recomendado):
#   python -m venv venv
#   source venv/bin/activate  # Linux/macOS
#   .\venv\Scripts\activate   # Windows
#   pip install -r requirements.txt

# ============================================
# DEPENDÊNCIA EXTERNA NECESSÁRIA
# ============================================
# FFmpeg: Deve estar instalado no sistema e disponível no PATH
#   - Necessário apenas para scripts de vídeo (otimizador-video.py, webm-mp4.py)
#   - Download: https://ffmpeg.org/download.html
#
# Instalação:
#   Windows: https://www.gyan.dev/ffmpeg/builds/ ou via chocolatey: choco install ffmpeg
#   Linux: sudo apt-get install ffmpeg (Ubuntu/Debian) ou sudo yum install ffmpeg (CentOS/RHEL)
#   macOS: brew install ffmpeg
#
# pngquant: Opcional - para compressão adicional de PNGs
#   Windows: https://pngquant.org/ ou via chocolatey: choco install pngquant
#   Linux: sudo apt-get install pngquant
#   macOS: brew install pngquant

# ============================================
# PROCESSAMENTO DE IMAGENS
# ============================================
# Pillow: Biblioteca para manipulação de imagens
#   - Usado em: otimizador-imagens.py, webp-to-jpg.py
Pillow>=10.2.0

# pillow-heif: Suporte para HEIC/HEIF (opcional)
#   - Usado em: otimizador-imagens.py (para converter HEIC/HEIF)
#   - Instalar apenas se precisar processar arquivos HEIC/HEIF
# pillow-heif>=0.13.0

# ============================================
# PROCESSAMENTO DE VÍDEOS E ANÁLISE DE IMAGENS
# ============================================
# OpenCV: Biblioteca para processamento de imagens e vídeos
#   - Usado em: validate-images.py, estabilizador-video.py
opencv-python>=4.12.0

# NumPy: Biblioteca para computação numérica
#   - Usado em: validate-images.py, estabilizador-video.py (análise de imagens)
numpy>=2.2.0

# ============================================
# UTILITÁRIOS
# ============================================
# tqdm: Barras de progresso em tempo real
#   - Usado em: todos os scripts
tqdm>=4.66.1

# python-dotenv: Carregamento de variáveis de ambiente
#   - Usado para configurações opcionais via .env
python-dotenv>=1.0.0

# requests: Biblioteca HTTP para requisições
#   - Usado para funcionalidades futuras (downloads, APIs, etc.)
requests>=2.31.0

# ============================================
# DEPENDÊNCIAS OPCIONAIS (descomente se necessário)
# ============================================
# Instale apenas se precisar das funcionalidades específicas:
#
# pillow-heif>=0.13.0  # Para suporte HEIC/HEIF (otimizador-imagens.py)
#   Instalar: pip install pillow-heif
#
# pytesseract>=0.3.10  # Para OCR (ocr-imagens.py - requer Tesseract instalado)
#   Instalar: pip install pytesseract
#   E instale Tesseract OCR no sistema:
#   - Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   - Linux: sudo apt-get install tesseract-ocr tesseract-ocr-por
#   - macOS: brew install tesseract tesseract-lang
#
# rembg>=2.0.50  # Para remoção de fundo (remover-fundo.py)
#   Instalar: pip install rembg
#   Nota: A primeira execução baixará o modelo (~170MB)
